Refactor: move request dir into invitation service
