+++ src/main/java/world/trecord/controller/feed/FeedController.java
@@ -11,12 +11,12 @@ import world.trecord.config.security.CurrentUser;
 import world.trecord.controller.ApiResponse;
 import world.trecord.service.feed.FeedService;
 import world.trecord.service.feed.request.FeedCreateRequest;
-import world.trecord.service.feed.request.FeedInviteRequest;
 import world.trecord.service.feed.request.FeedUpdateRequest;
 import world.trecord.service.feed.response.FeedCreateResponse;
 import world.trecord.service.feed.response.FeedInfoResponse;
 import world.trecord.service.feed.response.FeedListResponse;
 import world.trecord.service.invitation.InvitationService;
+import world.trecord.service.invitation.request.FeedInviteRequest;
 import world.trecord.service.users.UserContext;
 
 import java.util.Optional;
